# swagger_spec.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Voz Urbana",
        "description": "API para gerenciar usuários, petições, relatórios e imagens no Aplicativo e Pagina Web Voz Urbana!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Servidor Local"
        },
        {
            "url": "http://voz-urbana-api-cesargabriel26-cesar-gabriels-projects.vercel.app",
            "description": "Servidor De Testes"
        },
        {
            "url": "http://voz-urbana-api.vercel.app",
            "description": "Servidor De Produção"
        }
    ],
    "paths": {
        "/usuarios/create": {
            "post": {
                "tags": ["usuario"],
                "summary": "Criar um novo usuário",
                "description": "Cria um novo usuário com nome, email, senha, CPF e uma foto de perfil (opcional).",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "example": "senhaSegura123"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "pfp": {
                                        "type": "string",
                                        "example": "link_para_foto.jpg"
                                    }
                                },
                                "required": [
                                    "nome",
                                    "email",
                                    "senha",
                                    "cpf"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User added successfully"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios/login": {
            "post": {
                "tags": ["usuario"],
                "summary": "Login do usuário",
                "description": "Realiza o login de um usuário com email e senha.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "joao.silva@example.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "example": "senha_segura"
                                    }
                                },
                                "required": [
                                    "email",
                                    "senha"
                                ]
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Login realizado com sucesso",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "Usuario Encontrado"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "token_de_autenticacao"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "401": {
                            "description": "Senha incorreta",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "Senha Incorreta"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "404": {
                            "description": "Usuário não encontrado",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "Email ou Senha incorretos"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "500": {
                            "description": "Erro no servidor",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "error": {
                                                "type": "string",
                                                "example": "Detalhes do erro"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios/passwordCheck/{id}": {
            "post": {
                "tags": ["usuario"],
                "summary": "Verificar a senha do usuário",
                "description": "Verifica se a senha fornecida pelo usuário é igual à senha armazenada no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do usuário para o qual a senha será verificada.",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "senha": {
                                        "type": "string",
                                        "example": "senhaSegura123"
                                    }
                                },
                                "required": [
                                    "senha"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha verificada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Senha Correta"
                                        },
                                        "content": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Senha incorreta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Senha Incorreta"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuário não encontrado"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios/update/{id}": {
            "put": {
                "tags": ["usuario"],
                "summary": "Atualizar um usuário",
                "description": "Atualiza os dados de um usuário existente pelo ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do usuário a ser atualizado",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao.silva@example.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "example": "nova_senha_segura"
                                    },
                                    "pfp": {
                                        "type": "string",
                                        "example": "nova_url_da_foto_de_perfil"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    }
                                },
                                "required": [
                                    "nome",
                                    "email",
                                    "cpf"
                                ]
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Usuário atualizado com sucesso",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "User updated successfully"
                                            },
                                            "content": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "404": {
                            "description": "Usuário não encontrado",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "User not found"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "500": {
                            "description": "Erro no servidor",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "error": {
                                                "type": "string",
                                                "example": "Detalhes do erro"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios/delete/{id}": {
            "delete": {
                "tags": ["usuario"],
                "summary": "Deletar um usuário",
                "description": "Remove um usuário da base de dados pelo ID fornecido.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do usuário a ser deletado.",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "example": "joao@email.com"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "123.456.789-00"
                    },
                    "pfp": {
                        "type": "string",
                        "example": "link_para_foto.jpg"
                    }
                }
            },
...
        }
    }
}